moved http handle functions to own file
